DEVOPS-180: use startupProbe for DHIS2 slow start (#7)
* DEVOPS-180: use startupProbe for DHIS2 slow start

liveness and readiness probes will only run after the startup probe was
successful. This is why initialDelaySeconds has been set to their
default of 0. If the startupProbe fails the restartPolicy kicks in and
restarts the container in DHIS2 pod.

DHIS2 seems to start within 1-2min in the current configuration on
kubernetes. Therefore a total time of 130s has been chosen. Startup
times might vary more, be shorter most of the time. We can fine tune
them in the stack or on deploy to find the best configuration.

* chore: remove initialDelaySeconds

as liveness/readiness probes are only run after the startup probe succeeded
initialDelaySeconds is 0 by default kubernetes